Move ContextActivatingXyzListeners to util.selection package
