+++ org.eclipse.buildship.ui/src/main/java/org/eclipse/buildship/ui/view/task/UiContributionManager.java
@@ -17,8 +17,8 @@ import org.eclipse.buildship.ui.UiPluginConstants;
 import org.eclipse.buildship.ui.util.nodeselection.ActionEnablingSelectionChangedListener;
 import org.eclipse.buildship.ui.util.nodeselection.ActionShowingContextMenuListener;
 import org.eclipse.buildship.ui.util.nodeselection.SelectionSpecificAction;
+import org.eclipse.buildship.ui.util.selection.ContextActivatingViewPartListener;
 import org.eclipse.buildship.ui.view.CollapseTreeNodesAction;
-import org.eclipse.buildship.ui.view.ContextActivatingViewPartListener;
 import org.eclipse.buildship.ui.view.ExpandTreeNodesAction;
 import org.eclipse.buildship.ui.view.ShowFilterAction;
 import org.eclipse.core.resources.ResourcesPlugin;
